+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## Unreleased
++ Add optional timeout to all Airthings API calls.
 
 ## 1.1.0
 ### Features